Added PrimRootFinder class to set correct g and p values in the AliceBob class and to perform modpow calculations. Modified AliceBob to work with PrimRootFinder.

+++ KeyExchange.java
@@ -5,30 +5,30 @@ import java.util.concurrent.CyclicBarrier;
 public class KeyExchange{
     public static final Object lock = new Object();
     CyclicBarrier barrier = new CyclicBarrier(2);
-    private double p = -1;
-    private double g = -1;
-    private double alicePreKey = -1;
-    private double bobPrekey = -1;
+    private int p = -1;
+    private int g = -1;
+    private int alicePreKey = -1;
+    private int bobPrekey = -1;
 
 
-    public synchronized void setPG(double p, double g){
+    public synchronized void setPG(int p, int g){
         this.p = p;
         this.g = g;
     }
 
-    public double getP(){
+    public int getP(){
         return p;
     }
 
-    public double getG(){
+    public int getG(){
         return g;
     }
 
-    public double getAlicePreKey(){
+    public int getAlicePreKey(){
         return alicePreKey;
     }
 
-    public double getBobPrekey(){
+    public int getBobPrekey(){
         return bobPrekey;
     }
 
@@ -44,13 +44,14 @@ public class KeyExchange{
         return bobPrekey == -1;
     }
 
-    public synchronized void setAlicePreKey(double key){
+    public synchronized void setAlicePreKey(int key){
         alicePreKey = key;
     }
 
-    public synchronized void setBobPreKey(double key){
+    public synchronized void setBobPreKey(int key){
         bobPrekey = key;
     }
 
 
 }
+
